When /api/bots is called, a copy of the bot setting with null credentials is sent. Refactored *Credentials class into separate package
