+++ src/main/kotlin/ink/abb/pogo/scraper/Main.kt
@@ -16,6 +16,7 @@ import com.pokegoapi.auth.PtcCredentialProvider
 import com.pokegoapi.exceptions.LoginFailedException
 import com.pokegoapi.exceptions.RemoteServerException
 import com.pokegoapi.util.SystemTimeImpl
+import ink.abb.pogo.scraper.util.credentials.*
 import ink.abb.pogo.scraper.services.BotService
 import ink.abb.pogo.scraper.util.Log
 import okhttp3.OkHttpClient
