+++ src/main/kotlin/ink/abb/pogo/scraper/util/credentials/GoogleAutoCredentials.kt
@@ -1,0 +1,3 @@
+package ink.abb.pogo.scraper.util.credentials
+
+data class GoogleAutoCredentials(var username: String = "", var password: String = "") : Credentials
