+++ src/main/kotlin/ink/abb/pogo/scraper/util/credentials/GoogleCredentials.kt
@@ -1,0 +1,3 @@
+package ink.abb.pogo.scraper.util.credentials
+
+data class GoogleCredentials(var token: String = "") : Credentials
