+++ src/main/kotlin/ink/abb/pogo/scraper/util/credentials/PtcCredentials.kt
@@ -1,0 +1,3 @@
+package ink.abb.pogo.scraper.util.credentials
+
+data class PtcCredentials(val username: String = "", val password: String = "") : Credentials
